refactor :: relocate packages
