+++ app/src/main/java/com/teamnk/kimiljung/activity/StartActivity.kt
@@ -1,11 +1,10 @@
 package com.teamnk.kimiljung.activity
 
 import android.os.Bundle
-import androidx.databinding.DataBindingUtil
 import com.teamnk.kimiljung.R
+import com.teamnk.kimiljung.activity.auth.LoginActivity
 import com.teamnk.kimiljung.base.BaseActivity
 import com.teamnk.kimiljung.databinding.ActivityStartBinding
-import com.teamnk.kimiljung.util.initializeBinding
 import com.teamnk.kimiljung.util.startIntent
 
 class StartActivity : BaseActivity<ActivityStartBinding>(R.layout.activity_start) {
