--- app/src/main/java/com/teamnk/kimiljung/util/BindingUtil.kt
@@ -1,11 +1,0 @@
-package com.teamnk.kimiljung.util
-
-import androidx.databinding.ViewDataBinding
-import androidx.lifecycle.LifecycleOwner
-
-fun initializeBinding(binding: ViewDataBinding, lifecycleOwner: LifecycleOwner) {
-    binding.lifecycleOwner = lifecycleOwner
-
-
-
-}
