+++ app/src/main/res/layout/activity_register.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.RegisterActivity">
+        tools:context=".activity.auth.RegisterActivity">
 
         <TextView
             android:id="@+id/tv_register_mainTitle"
